requirements.txt
annotated-types==0.5.0
anyio==3.7.1
azure-core==1.29.4
azure-identity==1.14.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.4
defusedxml==0.7.1
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
httptools==0.6.0
idna==3.4
Jinja2==3.1.2
jira==3.5.2
MarkupSafe==2.1.3
msal==1.24.1
msal-extensions==1.0.0
numpy==1.26.0
oauthlib==3.2.2
packaging==23.2
pandas==2.1.1
portalocker==2.8.2
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
